requirements.txt
matplotlib~=3.4.1
scipy~=1.6.2
tqdm~=4.60.0
opencv-python~=4.5.1.48
pillow~=7.0.0
torch~=1.8.1
torchvision~=0.9.1
numpy~=1.20.2
scikit-image~=0.18.1
torchsummary~=1.5.1
tensorboardX~=2.2
PyQt5~=5.15.4
mysql-connector-python~=8.0.23